upload download script

+++ test/scripts/goparrun.sh
@@ -7,7 +7,7 @@ config=$3
 outputPath=$4
 
 sudo fusermount3 -u $mntPath
-rm -rf $mntPath/*
+sudo rm -rf $mntPath/*
 sudo rm -rf $tmpPath/*
 
 cnt=1
@@ -16,14 +16,16 @@ while IFS=, read -r thread count size; do
 
 	echo "Blobfuse2 | $cnt ($thread threads: $count files : $size MB) |"
 	./blobfuse2 mount $mntPath --config-file=$config &
+	if [ $? -ne 0 ]; then
+    	exit 1
+	fi
 	sleep 3
-	rm -rf $mntPath/*
+	ps -aux | grep blobfuse2
 
 	./test/scripts/pwrite.sh $thread $count $size $mntPath $outputPath $sed_line
 	sudo rm -rf $tmpPath/*
 	./test/scripts/pread.sh $thread $count $size $mntPath $outputPath $sed_line
 
-	rm -rf $mntPath/*
 	sudo fusermount3 -u $mntPath
 
 	(( cnt++ ))
